backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
psycopg2-binary==2.9.9
pytest==8.3.2
email-validator==2.2.0
